rewrote commands into typescript with a base class of command
