+++ package.json
@@ -5,6 +5,7 @@
     "main": "src/bot.js",
     "author": "Thomas Ortiz",
     "scripts": {
+        "run": "tsc; dist/bot.js",
         "sql": "heroku pg:psql -a pubg-stat-discord-bot",
         "sql-test": "heroku pg:psql -a pubg-stat-discord-bot-dev"
     },
