+++ src/models/season.ts
@@ -1,0 +1,5 @@
+export interface Season {
+    id: number;
+    name: string;
+    season: string;
+}
