+++ src/models/squadSize.ts
@@ -1,0 +1,5 @@
+export interface SquadSize {
+    id: number;
+    name: string;
+    size: number;
+}
