+++ src/services/sql-services/sql.squadSize.service.ts
@@ -1,6 +1,7 @@
 import * as logger from 'winston';
 import { CommonService as cs } from '../common.service';
 import { Pool } from 'pg';
+import { SquadSize } from '../../models/squadSize';
 
 
 let connectionString: string = cs.getEnvironmentVariable('DATABASE_URL');
@@ -15,13 +16,14 @@ pool.on('error', (err) => {
 
 
 export class SqlSqaudSizeService {
+
     /**
      *  Return all PUBG modes
-     * @returns {obj}: { id, name, size }
+     * @returns {Promise<any>}: { id, name, size }
      */
     static async getAllSquadSizes(): Promise<any> {
         return pool.query('select * from squad_sizes').then((res) => {
-            return res.rows;
+            return res.rows as SquadSize[];
         });
     }
 }
