+++ tsconfig.json
@@ -1,8 +1,7 @@
 {
     "compilerOptions": {
         "module": "commonjs",
-        "target": "es6",
-        "lib": ["es2015"],
+        "target": "es2017",
         //"noImplicitAny": true,
         "noUnusedLocals": true,
         "removeComments": true,
